moved envParamaters to resorces
