+++ src/core/resources/envParameters.ts
