+++ src/tests/ui/login.spec.ts
@@ -4,7 +4,7 @@ import { BasePage } from '../../business/page-objects/BasePage';
 import { UserDropDownMenu } from '../../business/page-objects/components/UserDropdownMenu';
 import config from '../../core/resources/config.json';
 import logger from '../../core/logger';
-import { RP_USERNAME, RP_PASSWORD } from '../../core/envParameters';
+import { RP_USERNAME, RP_PASSWORD } from '../../core/resources/envParameters';
 
 test.describe.parallel('Login and logout', () => {
   let basePage: BasePage;
